+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.1.9 (2023-10-20)
+
+### Changed
+* made the `priority` attribute of an `Annotation` non-Optional
+* multi token lookup now sets the `start_token` and `end_token` fields of an `Annotation`
+
+### Fixed
+* a bug with determnistic sort, when `Optional` fields were set
+
+
 ## 0.1.8 (2023-08-01)
 
 ### Added